Ease in with Pillar and FE types
