+++ dotcom-rendering/src/lib/decideDesign.ts
@@ -1,4 +1,5 @@
 import { ArticleDesign } from '@guardian/libs';
+import type { FEFormat } from 'types';
 
 /**
  * NOTE: Immersive Opinion pieces are not supported,
