+++ dotcom-rendering/src/lib/decideDisplay.ts
@@ -1,4 +1,5 @@
 import { ArticleDisplay } from '@guardian/libs';
+import type { FEFormat } from 'types';
 
 export const decideDisplay = ({
 	display,
