+++ dotcom-rendering/src/model/extract-nav.ts
@@ -1,4 +1,5 @@
 import { Pillar } from '@guardian/libs';
+import type { FELinkType, FENavType } from 'types';
 import type { EditionId } from '../lib/edition';
 import { findPillar } from './find-pillar';
