+++ dotcom-rendering/src/types/trails.ts
@@ -1,3 +1,4 @@
+import type { FEFormat } from 'types';
 import type { Branding } from './branding';
 import type { StarRating } from './content';
 import type { DCRFrontImage, DCRSnapType, DCRSupportingContent } from './front';
